requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-openai
flask-cors 